# apps/layer-gateway/openapi.yaml
...
openapi: 3.1.0
info:
  title: Hevlayer API
  description: 'Hevlayer-specific endpoints that extend the Turbopuffer API. These
    endpoints are not part of the Turbopuffer client — they are added by the Hevlayer
    proxy layer for document retrieval (cache), document cache operations, snapshot
    jobs, scans, result counts, and namespace metadata.

'
  version: 0.1.0
  contact:
    name: Hev
    url: https://github.com/hev/hevlayer
servers:
- url: http://localhost:8080
  description: Local development
- url: http://hevlayer.internal:8080
  description: In-cluster (Kubernetes)
security:
- bearerAuth: []
paths:
...
  "/v1/namespaces/{namespace}/metadata":
    get:
      operationId: getTurbopufferV1NamespaceMetadata
      x-hevlayer-store-support:
        turbopuffer: supported
        search: supported
      x-turbopuffer-key: GET /v1/namespaces/{namespace}/metadata
      summary: Get Turbopuffer v1 namespace metadata
      tags:
      - turbopuffer
      - namespaces
      parameters:
      - "$ref": "#/components/parameters/namespace"
      responses:
        '200':
          description: Namespace metadata
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/NamespaceMetadata"
        '502':
          description: Upstream error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
    patch:
      operationId: updateTurbopufferNamespaceMetadata
      x-hevlayer-store-support:
        turbopuffer: supported
        search: supported
      x-turbopuffer-key: PATCH /v1/namespaces/{namespace}/metadata
      summary: Update Turbopuffer namespace metadata
      description: Raw passthrough for upstream metadata configuration, currently
        used for namespace pinning.
      tags:
      - turbopuffer
      - namespaces
      parameters:
      - "$ref": "#/components/parameters/namespace"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TurbopufferMetadataPatch"
      responses:
        '200':
          description: Updated namespace metadata
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/NamespaceMetadata"
        '502':
          description: Upstream error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
...
  "/v2/query":
    post:
      operationId: query
      x-hevlayer-store-support:
        turbopuffer: supported
        search: supported
      summary: Federate one query across namespaces
      description: 'Runs one logical query against a caller-supplied set of namespaces
        and returns one fused `rows` list. The gateway fans out to the normal per-namespace
        query machinery, annotates every row with `$namespace` and `$rank`, and merges
        by distance for vector/semantic routes or by rank-interleave for text and
        hybrid routes. `namespaces: ["*"]` or an omitted `namespaces` field expands
        the authenticated key''s namespace allowlist at request time.

'
      tags:
      - query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/FederatedQueryRequest"
      responses:
        '200':
          description: Fused cross-namespace query response.
          headers:
            x-layer-stable-as-of:
              "$ref": "#/components/headers/LayerStableAsOf"
            x-layer-partial:
              "$ref": "#/components/headers/LayerPartial"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/FederatedQueryResponse"
        '422':
          description: 'Request validation error. Returned for an empty namespace
            set, `namespaces: ["*"]` mixed with other names, cursor usage, invalid
            fusion options, invalid underlying query shape, or strict vector merge
            with missing/mismatched embedding profiles.

'
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '403':
          description: A named namespace is outside the authenticated minted key's
            namespace grant.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '502':
          description: All namespace legs failed, or strict mode failed on one leg.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
...
  "/v2/namespaces/{namespace}/metadata":
    get:
      operationId: getNamespaceMetadata
      x-hevlayer-store-support:
        turbopuffer: supported
        search: supported
      x-turbopuffer-key: GET /v2/namespaces/{namespace}/metadata
      summary: Get namespace metadata
      description: 'Get metadata for a namespace including schema, row counts, byte
        sizes, and custom config.

'
      tags:
      - namespaces
      parameters:
      - "$ref": "#/components/parameters/namespace"
      responses:
        '200':
          description: Namespace metadata
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/NamespaceMetadata"
        '404':
          description: Namespace not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '502':
          description: Upstream error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
  "/v2/namespaces/{namespace}/init":
    post:
      operationId: initNamespace
      x-hevlayer-store-support:
        turbopuffer: supported
        search: unsupported
      summary: Initialize namespace metadata and shard backfill
      description: 'Persist Layer''s reserved namespace marker row, start or reattach
        to the embedded shard backfill task, and return the current init status. Scatter/gather
        remains inactive until `shard_lag_rows` reaches zero.

'
      tags:
      - namespaces
      parameters:
      - "$ref": "#/components/parameters/namespace"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/InitNamespaceRequest"
      responses:
        '200':
          description: Namespace init accepted or reattached
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/InitNamespaceResponse"
        '404':
          description: Namespace not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '409':
          description: Requested shard_count conflicts with the existing marker
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '422':
          description: Invalid init request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
        '502':
          description: Upstream error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
...
components:
  parameters:
    namespace:
      name: namespace
      in: path
      required: true
      schema:
        type: string
      description: Namespace name
...
  headers:
    LayerStableAsOf:
      description: Epoch-ms consistency watermark the read was served at.
      schema:
        type: integer
        format: int64
...
    LayerPartial:
      description: "`true` when a federated best-effort query returned successful
        namespace legs and reported one or more failed legs in `errors`.\n"
      schema:
        type: string
        enum:
        - 'true'
  schemas:
...
    TurbopufferMetadataPatch:
      type: object
      additionalProperties: true
      properties:
        pinning: {}
...
    IndexState:
      type: object
      description: 'Turbopuffer''s `index` metadata object, passed through verbatim.
        The raw upstream per-fetch signal of how caught up indexing is. Namespace
        list rows derive `is_stable` and `stable_as_of_ms` from this same metadata
        fetch. Additional upstream sub-fields ride through unmodified.

'
      additionalProperties: true
      properties:
        status:
          type: string
          enum:
          - updating
          - up-to-date
          description: '`"up-to-date"` when the upstream index is fully caught up;
            `"updating"` while writes are still being indexed.

'
        unindexed_bytes:
          type:
          - integer
          - 'null'
          format: int64
          description: 'Bytes written to the namespace''s write-ahead log but not
            yet indexed. Present only when `status == "updating"` (Turbopuffer omits
            it once caught up). Unindexed data is still searched by queries, so a
            non-zero value means "behind, but serving" — watch it fall to confirm
            indexing is draining rather than wedged.

'
...
    NamespaceMetadata:
      type: object
      required:
      - id
      - schema
      - approx_logical_bytes
      - approx_row_count
      - created_at
      - updated_at
      properties:
        id:
          type: string
        schema:
          type: object
          additionalProperties: true
        approx_logical_bytes:
          type: integer
        approx_row_count:
          type: integer
        created_at:
          type: string
          format: date-time
        last_write_at:
          type:
          - string
          - 'null'
          format: date-time
        updated_at:
          type: string
          format: date-time
        config:
          type: object
          additionalProperties: true
        index:
          "$ref": "#/components/schemas/IndexState"
        layer:
          "$ref": "#/components/schemas/NamespaceMetadataLayer"
    NamespaceMetadataLayer:
      type: object
      description: Gateway-computed namespace freshness and readiness signals.
      properties:
        stable_as_of:
          type:
          - integer
          - 'null'
          format: int64
          description: 'Epoch-ms watermark from the most recent stable consistency
            poll. Null before the watcher has observed a stable namespace.

'
        is_stable:
          type: boolean
          description: 'True when the most recent consistency poll observed `index.status
            == "up-to-date"`; false on cold start or while the upstream index is updating.

'
        indexed:
          type:
          - boolean
          - 'null'
          description: 'True when the latest snapshot''s indexed-row count is greater
            than or equal to the upstream namespace row count. False while vector
            rows are still missing from the indexed snapshot. Null for namespaces
            without a vector column, when the latest stored snapshot predates row-count
            capture, or when the snapshot lookup fails.

'
        index_lag_rows:
          type:
          - integer
          - 'null'
          format: int64
          description: 'Upstream namespace row count minus latest snapshot row count,
            clamped at zero. Null when `indexed` is null.

'
        schema_version:
          type: integer
          format: int64
          description: Layer reserved schema version stamped by namespace init.
        init_state:
          type: string
          enum:
          - running
          - ready
          description: Embedded namespace init drain state.
        init_lag_rows:
          type: integer
          format: int64
          description: Rows still missing init-owned reserved fields.
        shard_count:
          type:
          - integer
          - 'null'
          format: int64
          description: Persisted namespace shard count, or null for unsharded namespaces.
        shard_state:
          type: string
          enum:
          - unsharded
          - backfilling
          - ready
          description: Shard backfill readiness state.
        shard_lag_rows:
          type: integer
          format: int64
          description: Rows still missing `_hevlayer_shard`.
        scatter_gather_active:
          type: boolean
          description: True iff a shard marker exists and `shard_lag_rows` is zero.
    InitNamespaceRequest:
      type: object
      properties:
        schema_version:
          type: integer
          format: int64
          minimum: 1
          default: 1
          description: Layer reserved schema version to stamp on the namespace marker.
        shard_count:
          type: integer
          format: int64
          minimum: 1
          description: Number of `_hevlayer_shard` buckets to backfill.
    InitNamespaceResponse:
      type: object
      required:
      - namespace
      - layer
      properties:
        namespace:
          type: string
        layer:
          "$ref": "#/components/schemas/NamespaceMetadataLayer"
    QueryRequest:
      type: object
      description: |
        Vector query request body. Caller must supply exactly one of `vector` (raw query vector) or `nearest_to_id` (an array of document ids; the gateway resolves each vector via Aerospike, falling back to the namespace's configured VectorStore on miss, and averages them into a centroid query vector). The gateway may add a consistency watermark filter before forwarding to that store.
        `vector` is never returned in `attributes`. Listing `"vector"` in `include_attributes` causes the gateway to drop the field and set `x-layer-warning: vector_attribute_dropped` on the response.
      properties:
        vector:
          type: array
          items:
            type: number
          minItems: 1
          description: 'Query vector. Mutually exclusive with `nearest_to_id`; exactly
            one of the two is required.

'
        nearest_to_id:
          type: array
          items:
            type: string
          minItems: 1
          description: 'Document ids whose stored vectors are averaged component-wise
            into the centroid query vector. Mutually exclusive with `vector`; exactly
            one of the two is required, and the array must be non-empty. The gateway
            looks each vector up in Aerospike and falls back to the namespace''s configured
            VectorStore (with a cache backfill) on miss. Returns 404 naming any ids
            with no vector in either place.

'
        top_k:
          type: integer
          minimum: 1
          default: 10
          description: Number of nearest neighbors to return.
        filters:
          description: Turbopuffer filter expression.
        as_of:
          type: integer
          format: int64
          minimum: 0
          description: 'Epoch-ms temporal upper bound. Mutually exclusive with `between`.
            The gateway conjoins `_hevlayer_upserted_at <= as_of` with the caller
            filter before querying.

'
        between:
          type: array
          minItems: 2
          maxItems: 2
          items:
            type: integer
            format: int64
            minimum: 0
          description: 'Epoch-ms temporal window `[lo, hi]` interpreted as `lo < _hevlayer_upserted_at
            <= hi`. Mutually exclusive with `as_of` and requires `lo < hi`.

'
        include_attributes:
          oneOf:
          - type: boolean
          - type: array
            items:
              type: string
          description: 'Attributes to include in each result. `vector` is always stripped
            from the response; explicitly listing it triggers an `x-layer-warning`
            header.

'
        include_leg_breakdown:
          type: boolean
          default: false
          description: 'Opt in to per-row `$fused.legs` provenance on Layer-expanded
            fused responses (`HybridText`, and executed `Auto` routes that expand
            hybrid-text legs). Each leg entry names the effective leg (`bm25`, `fuzzy:<token>`,
            or `semantic`) and reports that row''s 1-based rank plus raw per-leg score/distance;
            `rank` and `score` are null when the row fell outside that leg''s `per_leg_limit`
            cut. Default false, so payload and unsharded query fan-out are unchanged
            for callers that do not ask.

'
        cursor:
          type: string
          description: 'Opaque pagination token returned by a prior response''s `next_cursor`.
            Resumes the same ranked query at the row that follows. Caller should treat
            as opaque. Native vector queries decode it into a `($dist > last OR ($dist
            == last AND id > last_id))` band filter. Layer-fused `HybridText` and
            executed `Auto` routes decode it into a bounded server-side re-fetch offset
            capped at 10,000 ranked rows.

'
        rank_by:
          type: array
          items: {}
          description: 'Rank expression tuple. With `top_k`, Layer handles the portable
            subset with cache, history, and stable-read behavior; native upstream
            `rank_by` bodies that omit `top_k` pass through unchanged. For search-store
            multivector namespaces, spell ANN over a vector bag as `["vectors", "ANN",
            [[...], [...]]]`. Two Layer-only operator spellings are intercepted: `["field",
            "HybridText", "input", {fuzziness?, rank_constant?, per_leg_limit?, threads?}?]`
            (hybrid text fusion, RFC 0022; `threads` is the same scatter/gather fan-out
            control as scans, defaulting from `Index.spec.scan.threads`) and `["field",
            "Auto", "input", {route?, vector?}?]` (query router, RFC 0044). When `rank_by`
            is present, `vector` and `nearest_to_id` must be absent.

'
    FederatedQueryRequest:
      allOf:
      - "$ref": "#/components/schemas/QueryRequest"
      - type: object
        properties:
          namespaces:
            type: array
            minItems: 1
            maxItems: 512
            items:
              type: string
              minLength: 1
            description: 'Explicit namespace set to query. `["*"]`, or omitting this
              field, expands the authenticated key''s namespace allowlist.

'
          strict:
            type: boolean
            default: false
            description: 'Fail the whole request on the first namespace leg error.
              Default false returns successful namespace legs and reports failed legs
              in `errors`.

'
          fusion:
            "$ref": "#/components/schemas/FederatedFusionOptions"
    FederatedFusionOptions:
      type: object
      properties:
        per_namespace_limit:
          type: integer
          minimum: 1
          description: 'Rows each namespace returns to the cross-namespace merge.
            Default is `clamp(2 * top_k, 10, 100)`.

'
        rank_constant:
          type: integer
          minimum: 1
          default: 60
          description: 'Reserved RRF constant echo. Inert on rank-interleave text
            routes.

'
    FederatedQueryResponse:
      type: object
      required:
      - rows
      - merge
      - namespaces
      properties:
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
          description: 'Fused rows. Every row includes `id`, `$namespace`, `$rank`,
            and route-native `$dist`/`$score` plus requested attributes when present.

'
        merge:
          type: object
          required:
          - method
          - route
          properties:
            method:
              type: string
              enum:
              - distance
              - rank-interleave
              - weighted-rrf
            route:
              type: string
              enum:
              - fused
              - dual-score
            downgraded_reason:
              type: string
              enum:
              - missing_embedding_profile
              - embedding_profile_mismatch
              description: 'Present when a vector route could not safely merge by
                raw distance and fell back to rank-interleave in best-effort mode.

'
        routing:
          "$ref": "#/components/schemas/RoutingEcho"
        hybrid:
          "$ref": "#/components/schemas/HybridEcho"
        namespaces:
          type: array
          items:
            "$ref": "#/components/schemas/FederatedNamespaceResult"
        errors:
          type: array
          items:
            "$ref": "#/components/schemas/FederatedNamespaceError"
...
    Error:
      type: object
      required:
      - error
      - message
      properties:
        error:
          type: string
        message:
          type: string
...